YAML support added
YAML is considered to be an easier to manage and understand language
when dealing with configurations. Now the deployer supports either JSON
or YAML as a means for configurations through dynamic detection and
thus enables both REST API via JSON concepts or native file in YAML as
desired for whatever integration mechanism is desired.